cleanup legacy pages+components no longer needed after api-v2
