--- src/components/Analysis/Functional/v2index.tsx
@@ -1,193 +1,0 @@
-import React, { useContext } from 'react';
-import { Link } from 'react-router-dom';
-
-import TabsForQueryParameter from 'components/UI/TabsForQueryParameter';
-import ExtLink from 'components/UI/ExtLink';
-import InfoBanner from 'components/UI/InfoBanner';
-
-import AnalysisContext from 'pages/Analysis/V2AnalysisContext';
-import { createSharedQueryParamContext } from 'hooks/queryParamState/useQueryParamState';
-// import PfamTab from 'components/Analysis/Functional/KO';
-import KOTab from 'components/Analysis/Functional/KO/v2Index';
-import PfamTab from 'components/Analysis/Functional/Pfam/v2Index';
-import useLegacyAnalysisKnownFiles from 'hooks/data/useLegacyAnalysisKnownFiles';
-import InterProTab from './InterPro/v2index';
-import GOTab from './GO/v2index';
-import { SharedTextQueryParam } from 'hooks/queryParamState/QueryParamStore/QueryParamContext';
-
-const PARAMETER_NAME = 'type';
-const PARAMETER_DEFAULT = 'interpro';
-const tabs = [
-  { label: 'InterPro', to: 'interpro' },
-  { label: 'GO Terms', to: 'go' },
-  { label: 'Pfam', to: 'pfam' },
-  { label: 'KO', to: 'ko' },
-];
-
-const { useType, withQueryParamProvider } = createSharedQueryParamContext({
-  type: SharedTextQueryParam(PARAMETER_DEFAULT),
-});
-
-const FunctionalAnalysis: React.FC = () => {
-  const { overviewData: data } = useContext(AnalysisContext);
-  const { resultsDir, interproPath, goPath } = useLegacyAnalysisKnownFiles();
-
-  const [type] = useType<string>();
-  const activeType = type || PARAMETER_DEFAULT;
-
-  if (!data) {
-    return <div>Loading...</div>; // or whatever loading component you prefer
-  }
-  // const accession = data.id;
-  // const version = Number(overviewData.attributes['pipeline-version']);
-  const versionStr = data.pipeline_version || '';
-  const version = parseFloat(versionStr.replace(/^V/i, ''));
-  const isLegacy = !Number.isNaN(version) && version < 6;
-  const longReadExperiment = data.experiment_type === 'LRASS';
-
-  if (version < 4.1) {
-    return (
-      <div className="vf-stack vf-stack--200">
-        <p>
-          Functional analysis files can be found from the{' '}
-          <a href={resultsDir} target="_blank" rel="noopener noreferrer">
-            results directory
-          </a>
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="vf-stack">
-      <p>
-        These charts present the functional analysis outputs of our pipeline,
-        which focus on{' '}
-        <ExtLink href="http://www.ebi.ac.uk/interpro" title="InterPro website">
-          InterPro
-        </ExtLink>
-        {version >= 5 && (
-          <>
-            ,{' '}
-            <ExtLink
-              href="https://www.ebi.ac.uk/interpro/entry/pfam/"
-              title="InterPro protein families (Pfam)"
-            >
-              Pfam
-            </ExtLink>
-            ,{' '}
-            <ExtLink
-              href="https://www.genome.jp/kegg/ko.html"
-              title="KEGG orthologue (KO) webpage"
-            >
-              KEGG orthologue
-            </ExtLink>
-          </>
-        )}{' '}
-        and{' '}
-        <ExtLink
-          href="https://www.uniprot.org/help/gene_ontology"
-          title="Gene Ontology webpage"
-        >
-          GO
-        </ExtLink>{' '}
-        term annotations. These summarise the functional content of the
-        sequences in the sample. The full set of results files can be found
-        under the <Link to="#download">Download</Link> tab.
-      </p>
-      {longReadExperiment && (
-        <InfoBanner title="Caution:" type="info">
-          The sequences in this sample are derived from long-read sequencing
-          technology. Gene-prediction on this sequence data can be problematic,
-          potentially resulting in fewer and/or truncated predictions.
-        </InfoBanner>
-      )}
-      <TabsForQueryParameter
-        tabs={version >= 5 ? tabs : tabs.slice(0, 4)}
-        queryParameter={PARAMETER_NAME}
-        defaultValue={PARAMETER_DEFAULT}
-      />
-      <div className="vf-tabs-content">
-        {activeType === 'interpro' && (
-          <InterProTab
-            isLegacy={isLegacy}
-            legacyFile={
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.url.endsWith('.summary.ips')
-              ) ||
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.alias.toLowerCase().includes('_summary.ips')
-              ) ||
-              interproPath
-            }
-          />
-        )}
-        {activeType === 'go' && (
-          <GOTab
-            isLegacy={isLegacy}
-            legacyFile={
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.url.endsWith('.summary.go')
-              ) ||
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.alias.toLowerCase().includes('_summary.go')
-              ) ||
-              goPath
-            }
-          />
-        )}
-        {activeType === 'ko' && (
-          <KOTab
-            isLegacy={isLegacy}
-            legacyFile={
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.url.endsWith('.summary.ko')
-              ) ||
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.alias.toLowerCase().includes('_summary.ko')
-              )
-            }
-          />
-        )}
-        {activeType === 'pfam' && (
-          <PfamTab
-            isLegacy={isLegacy}
-            legacyFile={
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.url.endsWith('.summary.pfam')
-              ) ||
-              data.downloads.find(
-                (f) =>
-                  (f.download_type === 'Functional analysis' ||
-                    f.download_group.includes('functional')) &&
-                  f.alias.toLowerCase().includes('_summary.pfam')
-              )
-            }
-          />
-        )}
-      </div>
-    </div>
-  );
-};
-export default withQueryParamProvider(FunctionalAnalysis);
