+++ src/pages/Analysis/index.tsx
@@ -1,115 +1,156 @@
 import React, { useMemo } from 'react';
 
-import useMGnifyData from '@/hooks/data/useMGnifyData';
-import { MGnifyDatum, MGnifyResponseObj } from '@/hooks/data/useData';
-import useURLAccession from '@/hooks/useURLAccession';
+import useURLAccession from 'hooks/useURLAccession';
 import Loading from 'components/UI/Loading';
 import FetchError from 'components/UI/FetchError';
-import Tabs from 'components/UI/Tabs';
+import { RouteTabs } from 'components/UI/Tabs';
 import Overview from 'components/Analysis/Overview';
-import QualityControl from 'components/Analysis/QualityControl';
-import ContigViewer from 'components/Analysis/ContigViewer';
-import TaxonomySubpage from 'components/Analysis/AmpliconTaxonomy';
+import QualityControl from 'components/Analysis/QualityControl/v2index';
+import LegacyQualityControl from 'components/Analysis/QualityControl/index';
+import LegacyContigViewer from 'components/Analysis/ContigViewer/index';
+import ContigViewer from 'components/Analysis/ContigViewer/v2index';
 import FunctionalSubpage from 'components/Analysis/Functional';
 import PathwaysSubpage from 'components/Analysis/Pathways';
-import RouteForHash from 'components/Nav/RouteForHash';
-import { Link } from 'react-router-dom';
-import Downloads from 'components/Downloads';
-import Abundance from 'components/Analysis/Abundance';
-import AnalysisContext from './AnalysisContext';
+import Downloads from 'components/Downloads/v2index';
+import V2AnalysisContext, {
+  AnalysisContextType,
+} from 'pages/Analysis/V2AnalysisContext';
+import useAnalysisDetail from 'hooks/data/useAnalysisDetail';
+import { AnalysisDetail } from '@/interfaces';
+import Asv from 'components/Asv';
+import WGSTaxonomy from 'components/Analysis/WGSTaxonomy';
+import AmpliconTaxonomy from 'components/Analysis/AmpliconTaxonomy';
+import LegacyTaxonomy from 'components/Analysis/Taxonomy/LegacyTaxonomy';
+import Breadcrumbs from 'components/Nav/Breadcrumbs';
+import { Navigate, Route, Routes } from 'react-router-dom';
 
-const hasAbundance = (
-  includes: { attributes?: { 'group-type'?: string } }[]
+const isAssembly = (
+  experimentType: AnalysisDetail['experiment_type']
+): boolean => experimentType.toLowerCase().endsWith('assembly');
+
+const isAmplicon = (
+  experimentType: AnalysisDetail['experiment_type']
 ): boolean => {
-  return includes.some(
-    (included) => included?.attributes?.['group-type'] === 'Statistics'
-  );
+  return experimentType.toLowerCase() === 'amplicon';
 };
-
-const isAssembly = (data: MGnifyDatum): boolean =>
-  ['assembly', 'hybrid_assembly'].includes(
-    data.attributes['experiment-type'] as string
-  );
-const isAtleastVersion5 = (data: MGnifyDatum): boolean =>
-  Number(data.attributes['pipeline-version']) >= 5;
-const isNotAmplicon = (data: MGnifyDatum): boolean =>
-  data.attributes['experiment-type'] !== 'amplicon';
 
 const AnalysisPage: React.FC = () => {
   const accession = useURLAccession();
-  const { data, loading, error } = useMGnifyData(`analyses/${accession}`, {
-    include: 'downloads',
-  });
+  const { data, loading, error } = useAnalysisDetail(accession);
+
   const value = useMemo(
-    () => ({ overviewData: data?.data, included: data?.included ?? {} }),
-    [data]
+    () => ({ overviewData: data } as AnalysisContextType),
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [data?.accession]
   );
-  if (loading || !accession) return <Loading size="large" />;
+
+  const thisIsAssembly = useMemo(
+    () => data?.experiment_type && isAssembly(data.experiment_type),
+    [data?.experiment_type]
+  );
+  const thisIsAmplicon = useMemo(
+    () => data?.experiment_type && isAmplicon(data.experiment_type),
+    [data?.experiment_type]
+  );
+
+  const thisIsV6Family = useMemo(
+    () => data?.pipeline_version?.startsWith('V6'),
+    [data?.pipeline_version]
+  );
+
+  const pipelineVersion = useMemo(
+    () =>
+      data?.pipeline_version
+        ? parseFloat(data.pipeline_version.replace('V', ''))
+        : 0,
+    [data?.pipeline_version]
+  );
+
+  const tabs = useMemo(() => {
+    if (!data?.accession)
+      return [] as { label: string | React.ElementType; to: string }[];
+    return [
+      { label: 'Overview', to: 'overview' },
+      { label: 'Quality control', to: 'qc' },
+      { label: 'Taxonomy', to: 'taxonomic' },
+      thisIsAssembly
+        ? { label: 'Functional analysis', to: 'functional' }
+        : null,
+      thisIsAssembly && pipelineVersion >= 5
+        ? { label: 'Pathways/Systems', to: 'path-systems' }
+        : null,
+      thisIsAssembly && pipelineVersion >= 5
+        ? { label: 'Contig Viewer', to: 'contigs-viewer' }
+        : null,
+      thisIsAmplicon && thisIsV6Family ? { label: 'ASV', to: 'asv' } : null,
+      { label: 'Downloads', to: 'download' },
+    ].filter(Boolean) as { label: string | React.ElementType; to: string }[];
+  }, [
+    data?.accession,
+    thisIsAssembly,
+    thisIsAmplicon,
+    thisIsV6Family,
+    pipelineVersion,
+  ]);
+
+  const breadcrumbs = useMemo(() => {
+    if (!data?.accession) return [] as Array<{ label: string; url?: string }>;
+    return [
+      { label: 'Home', url: '/' },
+      { label: 'Studies', url: '/browse/studies' },
+      { label: data.study_accession, url: '/studies/' + data.study_accession },
+      { label: data.accession },
+    ];
+  }, [data?.accession, data?.study_accession]);
+
+  if (loading) return <Loading size="large" />;
   if (error) return <FetchError error={error} />;
   if (!data) return <Loading />;
-  const { data: analysisData, included } = data as MGnifyResponseObj;
 
-  const tabs = [
-    { label: 'Overview', to: '#overview' },
-    { label: 'Quality control', to: '#qc' },
-    { label: 'Taxonomic analysis', to: '#taxonomic' },
-    isNotAmplicon(analysisData)
-      ? { label: 'Functional analysis', to: '#functional' }
-      : null,
-    included && hasAbundance(included)
-      ? { label: 'Abundance and comparison', to: '#abundance' }
-      : null,
-    isAssembly(analysisData) && isAtleastVersion5(analysisData)
-      ? { label: 'Pathways/Systems', to: '#path-systems' }
-      : null,
-    isAssembly(analysisData) && isAtleastVersion5(analysisData)
-      ? { label: 'Contig Viewer', to: '#contigs-viewer' }
-      : null,
-    { label: 'Download', to: '#download' },
-  ].filter(Boolean) as { label: string; to: string }[];
-  const linkToOtherAnalyses = isAssembly(analysisData)
-    ? `/assemblies/${analysisData?.relationships?.assembly?.data?.id}`
-    : `/runs/${analysisData?.relationships?.run?.data?.id}`;
   return (
     <section className="vf-content">
+      <Breadcrumbs links={breadcrumbs} />
       <h2>Analysis {accession}</h2>
-      {linkToOtherAnalyses && (
-        <h4>
-          <Link to={linkToOtherAnalyses}>
-            <i className="icon icon-common icon-arrow-circle-left" /> Other
-            Analyses
-          </Link>
-        </h4>
-      )}
-      <Tabs tabs={tabs} />
+      <RouteTabs tabs={tabs} />
       <section className="vf-grid">
         <div className="vf-stack vf-stack--200">
-          <AnalysisContext.Provider value={value}>
-            <RouteForHash hash="#overview" isDefault>
-              <Overview />
-            </RouteForHash>
-            <RouteForHash hash="#qc">
-              <QualityControl />
-            </RouteForHash>
-            <RouteForHash hash="#contigs-viewer">
-              <ContigViewer />
-            </RouteForHash>
-            <RouteForHash hash="#taxonomic">
-              <TaxonomySubpage accession={accession} />
-            </RouteForHash>
-            <RouteForHash hash="#functional">
-              <FunctionalSubpage />
-            </RouteForHash>
-            <RouteForHash hash="#abundance">
-              <Abundance />
-            </RouteForHash>
-            <RouteForHash hash="#path-systems">
-              <PathwaysSubpage />
-            </RouteForHash>
-            <RouteForHash hash="#download">
-              <Downloads endpoint="analyses" accession={accession} />
-            </RouteForHash>
-          </AnalysisContext.Provider>
+          <V2AnalysisContext.Provider value={value}>
+            <Routes>
+              <Route index element={<Navigate to="overview" replace />} />
+              <Route path="overview" element={<Overview />} />
+              <Route
+                path="qc"
+                element={
+                  thisIsV6Family ? <QualityControl /> : <LegacyQualityControl />
+                }
+              />
+              <Route path="asv" element={<Asv />} />
+              <Route
+                path="contigs-viewer/*"
+                element={
+                  thisIsV6Family ? <ContigViewer /> : <LegacyContigViewer />
+                }
+              />
+              <Route
+                path="taxonomic"
+                element={
+                  // eslint-disable-next-line no-nested-ternary
+                  thisIsV6Family ? (
+                    isAmplicon(data.experiment_type) ? (
+                      <AmpliconTaxonomy />
+                    ) : (
+                      <WGSTaxonomy />
+                    )
+                  ) : (
+                    <LegacyTaxonomy />
+                  )
+                }
+              />
+              <Route path="functional" element={<FunctionalSubpage />} />
+              <Route path="path-systems" element={<PathwaysSubpage />} />
+              <Route path="download" element={<Downloads />} />
+            </Routes>
+          </V2AnalysisContext.Provider>
         </div>
       </section>
     </section>
