+++ src/pages/Assembly/index.tsx
@@ -1,94 +1,210 @@
-import React from 'react';
+import React, { useState, useEffect, useMemo, useCallback } from 'react';
 
-import useMGnifyData from '@/hooks/data/useMGnifyData';
-import { MGnifyResponseObj } from '@/hooks/data/useData';
-import useURLAccession from '@/hooks/useURLAccession';
+import axios from 'utils/protectedAxios';
+import useURLAccession from 'hooks/useURLAccession';
 import Loading from 'components/UI/Loading';
 import FetchError from 'components/UI/FetchError';
 import Box from 'components/UI/Box';
-import KeyValueList, { KeyValueItemsList } from 'components/UI/KeyValueList';
-import ExtLink from 'components/UI/ExtLink';
-import { Link } from 'react-router-dom';
 import AssociatedAnalyses from 'components/Analysis/Analyses';
-import { ENA_VIEW_URL } from '@/utils/urls';
 import ExtraAnnotations from 'components/ExtraAnnotations';
-import { createSharedQueryParamContextForTable } from 'hooks/queryParamState/useQueryParamState';
+import Tabs from 'components/UI/Tabs';
+import RouteForHash from 'components/Nav/RouteForHash';
+import config from 'utils/config';
+import InfoBanner from 'components/UI/InfoBanner';
+import DerivedGenomes from 'components/Assembly/DerivedGenomes';
+import AdditionalContainedGenomes from 'components/Assembly/AdditionalContainedGenomes';
+import V2AssemblyContext from './context';
+import { ErrorFromFetch, ErrorTypes } from 'hooks/data/useData';
 
-const { withQueryParamProvider } =
-  createSharedQueryParamContextForTable('analyses');
+// ---------- Types ----------
+
+type GenomeLink = {
+  genome: {
+    accession: string;
+    taxon_lineage: string;
+    catalogue_id: string | number;
+    catalogue_version: string | number;
+  };
+  species_rep: string;
+  mag_accession?: string;
+  updated_at?: string;
+};
+
+// Normalised data type for this page
+type AssemblyLinksData = {
+  accession: string;
+  genome_links: GenomeLink[];
+  run_accession?: React.ReactNode;
+  sample_accession?: string;
+};
+
+type GenomeLinksResponse = {
+  count?: number;
+  items: GenomeLink[];
+};
+
+const Genomes: React.FC = () => {
+  const { genomeLinksLastUpdated, assemblyData } =
+    React.useContext(V2AssemblyContext);
+
+  const [showBanner, setShowBanner] = React.useState(true);
+
+  const derivedGenomesAccessions = useMemo(
+    () => assemblyData?.genome_links?.map((gl) => gl.genome.accession) || [],
+    [assemblyData]
+  );
+
+  return (
+    <>
+      {showBanner && (
+        <InfoBanner
+          type="success"
+          dismissible
+          dismissAriaLabel="Dismiss last updated banner"
+          onDismiss={() => setShowBanner(false)}
+        >
+          <p className="vf-banner__text">
+            Last updated on:{' '}
+            {genomeLinksLastUpdated
+              ? new Date(genomeLinksLastUpdated).toLocaleString()
+              : 'N/A'}
+          </p>
+        </InfoBanner>
+      )}
+
+      <DerivedGenomes />
+      <AdditionalContainedGenomes derivedGenomes={derivedGenomesAccessions} />
+    </>
+  );
+};
+
+const Analyses: React.FC = () => {
+  return (
+    <Box label="Associated analyses">
+      <AssociatedAnalyses rootEndpoint="assemblies" />
+    </Box>
+  );
+};
+
+const AdditionalAnalyses: React.FC = () => {
+  return (
+    <Box label="Additional analyses">
+      <p>
+        Additional annotations produced by workflows run outside the scope of
+        MGnify&apos;s versioned pipelines.
+      </p>
+      <ExtraAnnotations namespace="assemblies" />
+    </Box>
+  );
+};
 
 const AssemblyPage: React.FC = () => {
   const accession = useURLAccession();
-  const { data, loading, error } = useMGnifyData(`assemblies/${accession}`);
-  if (loading) return <Loading size="large" />;
-  if (error) return <FetchError error={error} />;
-  if (!data) return <Loading />;
-  const { data: assemblyData } = data as MGnifyResponseObj;
 
-  const details = [
-    {
-      key: 'Sample',
-      value: assemblyData?.relationships?.samples?.data?.length
-        ? () => (
-            <>
-              {assemblyData.relationships.samples?.data.map((sample) => (
-                <Link to={`/samples/${sample.id}`} key={sample.id as string}>
-                  {sample.id}{' '}
-                </Link>
-              ))}
-            </>
-          )
+  const [data, setData] = useState<AssemblyLinksData | null>(null);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<{
+    error: unknown;
+    type: ErrorTypes;
+  } | null>(null);
+
+  const [genomeLinksLastUpdated, setGenomeLinksLastUpdated] = useState<
+    string | null
+  >(null);
+
+  const fetchData = useCallback(async () => {
+    setLoading(true);
+    setError(null);
+    if (!accession) return;
+
+    try {
+      const response = await axios.get<GenomeLinksResponse>(
+        `${config.api_v2}assemblies/${accession}/genome-links`
+      );
+      const { items } = response.data;
+      setGenomeLinksLastUpdated(items?.[0]?.updated_at ?? null);
+      const normalised: AssemblyLinksData = {
+        accession,
+        genome_links: items.map((it) => ({
+          genome: it.genome,
+          species_rep: it.species_rep,
+          mag_accession: it.mag_accession,
+          updated_at: it.updated_at,
+        })),
+        run_accession: undefined,
+        sample_accession: '',
+      };
+
+      setData(normalised);
+    } catch (err) {
+      setError({
+        error: err,
+        type: ErrorTypes.FetchError,
+      });
+    } finally {
+      setLoading(false);
+    }
+  }, [accession]);
+
+  useEffect(() => {
+    fetchData();
+  }, [fetchData]);
+
+  const assemblyData = data;
+
+  const ctxValue = useMemo(
+    () => ({
+      assemblyData: assemblyData
+        ? {
+            accession: assemblyData.accession,
+            run_accession: assemblyData.run_accession,
+            genome_links: assemblyData.genome_links,
+            sample_accession: assemblyData.sample_accession ?? '',
+          }
         : null,
-    },
-    {
-      key: 'Runs',
-      value: assemblyData?.relationships?.runs?.data?.length
-        ? () => (
-            <>
-              {assemblyData.relationships.runs?.data.map((run) => (
-                <Link to={`/runs/${run.id}`} key={run.id as string}>
-                  {run.id}{' '}
-                </Link>
-              ))}
-            </>
-          )
-        : null,
-    },
-    {
-      key: 'ENA accession',
-      value: () => (
-        <ExtLink href={`${ENA_VIEW_URL}${assemblyData?.id}`}>
-          {assemblyData?.id}
-        </ExtLink>
-      ),
-    },
-    {
-      key: 'Legacy accession',
-      value: assemblyData.attributes['legacy-accession'] as string,
-    },
-  ].filter(({ value }) => Boolean(value));
+      loading,
+      error,
+      refetch: fetchData,
+      genomeLinksLastUpdated,
+    }),
+    [assemblyData, loading, error, fetchData, genomeLinksLastUpdated]
+  );
+
+  if (loading) return <Loading size="large" />;
+  if (error) return <FetchError error={error as ErrorFromFetch} />;
+  if (!data) return <Loading />;
+
+  const tabs = [
+    { label: 'Analyses', to: '#analyses' },
+    { label: 'Genomes', to: '#genomes' },
+    // TODO: Re-enable when we have assembly extra annotations API
+    // { label: 'Additional Analyses', to: '#additional-analyses' },
+  ];
+
   return (
     <section className="vf-content">
-      <h2>Assembly: {assemblyData?.id || ''}</h2>
+      <h2>Assembly: {assemblyData?.accession || ''}</h2>
+      <Tabs tabs={tabs} />
+
       <section className="vf-grid">
         <div className="vf-stack vf-stack--200">
-          <Box label="Description">
-            <KeyValueList list={details as KeyValueItemsList} />
-          </Box>
-          <Box label="Associated analyses">
-            <AssociatedAnalyses rootEndpoint="assemblies" />
-          </Box>
-          <Box label="Additional analyses">
-            <p>
-              Additional annotations produced by workflows run outside the scope
-              of MGnify’s versioned pipelines.
-            </p>
-            <ExtraAnnotations namespace="assemblies" />
-          </Box>
+          <V2AssemblyContext.Provider value={ctxValue}>
+            <RouteForHash hash="#analyses" isDefault>
+              <Analyses />
+            </RouteForHash>
+
+            <RouteForHash hash="#genomes">
+              <Genomes />
+            </RouteForHash>
+
+            <RouteForHash hash="#additional-analyses">
+              <AdditionalAnalyses />
+            </RouteForHash>
+          </V2AssemblyContext.Provider>
         </div>
       </section>
     </section>
   );
 };
 
-export default withQueryParamProvider(AssemblyPage);
+export default AssemblyPage;
